Don't show settings when logged out
We were getting a NPE when logged out, since the user object is
undefined. We could put settings for logged-out users in the session...
but it seems far easier for now to simply say that settings don't appear
in the UI when logged out.

(We'll also need to handle this case later, when consuming settings.)